Cambio en el tipo de dato del salario

+++ database/migrations/2024_10_19_001839_create_empleados_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('posicion');
             $table->foreignId('departamento_id')->constrained()->onDelete('cascade');
             $table->date('fecha_contratacion');
-            $table->decimal('salario', 8, 2);
+            $table->integer('salario');
             $table->timestamps();
         });
     }
